Added niching test cases.
Added some unit tests, many more need to be added though. The NichePSO
is rather complex and as a result the test cases need to be crafted very
carefully.

Corrected some type casts and added a new method to StatUtils.

+++ src/main/java/net/sourceforge/cilib/pso/niching/StandardAbsorptionStrategy.java
@@ -4,17 +4,17 @@
  * Department of Computer Science
  * University of Pretoria
  * South Africa
- * 
+ *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation; either version 2 of the License, or
  * (at your option) any later version.
- * 
+ *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU General Public License
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
@@ -50,7 +50,7 @@ public class StandardAbsorptionStrategy implements AbsorptionStrategy {
             RadiusVisitor radiusVisitor = new RadiusVisitor();
             pba.accept(radiusVisitor);
 
-            double radius = ((Double) radiusVisitor.getResult()).doubleValue();
+            double radius = radiusVisitor.getResult().doubleValue();
 
             for (Entity entity : algorithm.getTopology()) {
                 double distance = distanceMeasure.distance(entity.getCandidateSolution(), algorithm.getTopology().getBestEntity().getCandidateSolution());
@@ -64,8 +64,6 @@ public class StandardAbsorptionStrategy implements AbsorptionStrategy {
                 }
             }
         }
-
-//        throw new UnsupportedOperationException("Not supported yet.");
     }
 
 }
